# docs/swagger.yaml
definitions:
  models.AddUser:
    properties:
      email:
        type: string
      full_name:
        type: string
      handphone:
        type: string
      is_admin:
        type: boolean
      password:
        type: string
    type: object
...
  models.ResponseModelList:
    properties:
      data:
        type: object
      last_page:
        type: integer
      message:
        type: string
      page:
        type: integer
      total:
        type: integer
    type: object
...
  models.TblTransaksi:
    properties:
      barang_cd:
        type: string
      created_on:
        type: integer
      deleted_on:
        type: integer
      harga_jual:
        type: number
      harga_modal:
        type: number
      id:
        type: integer
      id_created:
        type: integer
      keuntungan:
        type: number
      modified_on:
        type: integer
      pembeli:
        type: string
    type: object
  tool.ResponseModel:
    properties:
      data:
        type: object
      message:
        description: Code int         `json:"code"`
        type: string
    type: object
info:
  contact:
    email: nuryantofattih@gmail.com
    name: Nuryanto
    url: https://www.linkedin.com/in/nuryanto-1b2721156/
  description: Backend REST API for golang nuryanto2121
  license: {}
  title: Starter
  version: "1.0"
paths:
...
  /api/transaksi:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tool.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Delete Transaksi
      tags:
      - Transaksi
    get:
      parameters:
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      - description: PerPage
        in: query
        name: perpage
        required: true
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      - description: InitSearch
        in: query
        name: initsearch
        type: string
      - description: SortField
        in: query
        name: sortfield
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseModelList'
      security:
      - ApiKeyAuth: []
      summary: GetList Transaksi
      tags:
      - Transaksi
    post:
      parameters:
      - description: 'req param #changes are possible to adjust the form of the registration
          form from frontend'
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.TblTransaksi'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tool.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Add Transaksi
      tags:
      - Transaksi
  /api/transaksi/{id}:
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tool.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: GetById
      tags:
      - Transaksi
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: 'req param #changes are possible to adjust the form of the registration
          form from frontend'
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.TblTransaksi'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tool.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Update Transaksi
      tags:
      - Transaksi
  /api/user:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tool.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Delete User
      tags:
      - User
    get:
      parameters:
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      - description: PerPage
        in: query
        name: perpage
        required: true
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      - description: InitSearch
        in: query
        name: initsearch
        type: string
      - description: SortField
        in: query
        name: sortfield
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseModelList'
      security:
      - ApiKeyAuth: []
      summary: GetList User
      tags:
      - User
    post:
      parameters:
      - description: 'req param #changes are possible to adjust the form of the registration
          form from frontend'
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.AddUser'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tool.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Add User
      tags:
      - User
  /api/user/{id}:
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tool.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: GetById
      tags:
      - User
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: 'req param #changes are possible to adjust the form of the registration
          form from frontend'
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.AddUser'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tool.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Update User
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"